Move batch_download function to proper file
